refactor: 将评论数量格式化逻辑提取为工具函数并在歌单页面复用
